Redesign philosophy.
- Merge Schedule and TimeStream;
- Make Cooldown just an ActivityProcess;
- Expose RealThing.aspects as a read-only map.

+++ src/main/kotlin/org/tendiwa/time/ActivityProcess.kt
@@ -1,6 +1,11 @@
 package org.tendiwa.time
 
+/**
+ * @param duration Duration of the process in time quantums.
+ * @param result What happens when the process finishes, i.e. when time equal
+ * to [duration] passes since the beginning of the process.
+ */
 data class ActivityProcess(
-    val duration: Double,
+    val duration: Int,
     val result: ActivityResult
 )
